requirements.txt
PyQt6==6.7.1
pymodbus==3.6.5
paho-mqtt==2.1.0
asyncua==1.1.5
pyqtgraph==0.13.7
pandas==2.2.2